pyproject.toml
[tool.poetry]
name = "listening-station"
version = "0.1.0"
description = ""
authors = ["Dan Cooper <iamdanielcooper@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
spotipy = "^2.23.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
